Implement folder selector in edit bookmarks to move a bookmark.
The folders are sorted by my recently modified, with the Toolbar,
Menu, and Other Bookmarks pinned to the top.

The html select dropdown assumes that the selected attribute does not
move when a new property is selected. Firefox behaves like this,
however I do not know if all browsers behave this way. If they don't,
I can easily populate the data-parent-id attribute next to data-id in
the form.

We hide the root folder, and the mobile folder. I may display the
mobile folder later on once I figure out what determines it's
visibility. It's not visible for me when adding bookmarks since I am
not loggen into a Mozilla account.

Just remembered that firefox icons+css won't work in Chrome. So
bundling the arrow-dropdown-12.svg instead of using the chrome:// url.