Implement bookmark draganddrop move.
The existing place-entry div is either moved to the new position.
If dragged from a readonly group, we check to see if the bookmark is
visible in a pinned folder, and move that div. If not, we create a new
place entry.

We also check to see if the bookmark dropped on is still in the pinned
folder, and if not perform a full refresh.