Add function to delete a cached favicon
I noticed that when using BreezeDark ColorScheme in KDE caused Firefox
to use a Dark Mode, which told GitHub to use a favicon with a
hardcoded white fill.

I'll need to think about how I should detect this and either modify
the favicons, or if I should style the NewTab page with a dark theme.

Either way, I'll need a way to clear the favicon cache for when the
user switches between a Dark and Light ColorScheme in KDE.