Refactor most icons to be embedded SVGs in the DOM to theme them.
Firefox about:newtab uses fill="context-fill" in the SVGs, however
that is only supported on internal chrome:// URLs without an
about:config toggle.
* https://docs.w3cub.com/css/-moz-context-properties.html

So we need to embed the svg in the DOM to style it with CSS.
An alternative is to mask the bg-color and bg-image but that means the
button can't have a bg-color on hover which is more annoying.
* https://stackoverflow.com/questions/51395179/svg-fill-with-css-variables